converting totalprice to double in shoppingcart model

+++ Group1/Controllers/ShoppingcartController.cs
@@ -76,7 +76,7 @@ namespace Group1.Web.Controllers
 
             //calculate total price only if the cart contains data
             if (cart != null)
-                shoppingCart.TotalPrice =(decimal) shoppingCart.productlist.Sum(x => x.Product.Price * x.Quantity);
+                shoppingCart.TotalPrice =(double) shoppingCart.productlist.Sum(x => x.Product.Price * x.Quantity);
 
             return View(shoppingCart);
         }
